// Client/src/lib/utils/protos/adrBinaProto.ts
export interface AdrBinaResponse {
  isSuccess: boolean;
  message: string;
  statusCode: number;
  data: AdrBinaProto[];
}

export interface AdrBinaProto {
  id: number;
  kodu: string;
  siteAdi: string;
  adi: string;
  binaKatSayisi: number;
  daireSayisi: number;
  isyeriSayisi: number;
  yukseklik: number;
  wkb: string;
}

function createBaseAdrBinaResponse(): AdrBinaResponse {
  return { isSuccess: false, message: "", statusCode: 0, data: [] };
}

export const AdrBinaResponse: MessageFns<AdrBinaResponse> = {
  encode(message: AdrBinaResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.isSuccess !== false) {
      writer.uint32(8).bool(message.isSuccess);
    }
    if (message.message !== "") {
      writer.uint32(18).string(message.message);
    }
    if (message.statusCode !== 0) {
      writer.uint32(24).int32(message.statusCode);
    }
    for (const v of message.data) {
      AdrBinaProto.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AdrBinaResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAdrBinaResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.isSuccess = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.message = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.statusCode = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.data.push(AdrBinaProto.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AdrBinaResponse {
    return {
      isSuccess: isSet(object.isSuccess) ? globalThis.Boolean(object.isSuccess) : false,
      message: isSet(object.message) ? globalThis.String(object.message) : "",
      statusCode: isSet(object.statusCode) ? globalThis.Number(object.statusCode) : 0,
      data: globalThis.Array.isArray(object?.data) ? object.data.map((e: any) => AdrBinaProto.fromJSON(e)) : [],
    };
  },

  toJSON(message: AdrBinaResponse): unknown {
    const obj: any = {};
    if (message.isSuccess !== false) {
      obj.isSuccess = message.isSuccess;
    }
    if (message.message !== "") {
      obj.message = message.message;
    }
    if (message.statusCode !== 0) {
      obj.statusCode = Math.round(message.statusCode);
    }
    if (message.data?.length) {
      obj.data = message.data.map((e) => AdrBinaProto.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AdrBinaResponse>, I>>(base?: I): AdrBinaResponse {
    return AdrBinaResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AdrBinaResponse>, I>>(object: I): AdrBinaResponse {
    const message = createBaseAdrBinaResponse();
    message.isSuccess = object.isSuccess ?? false;
    message.message = object.message ?? "";
    message.statusCode = object.statusCode ?? 0;
    message.data = object.data?.map((e) => AdrBinaProto.fromPartial(e)) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}