// Client/src/lib/utils/protos/buildingsProto.ts
export interface BuildingsResponse {
  isSuccess: boolean;
  message: string;
  statusCode: string;
  data: BuildingsProto[];
}

export interface BuildingsProto {
  id: number;
  kodu: string;
  siteAdi: string;
  adi: string;
  binaKatSayisi: number;
  daireSayisi: number;
  isyeriSayisi: number;
  yukseklik: number;
  wkb: string;
}

function createBaseBuildingsResponse(): BuildingsResponse {
  return { isSuccess: false, message: "", statusCode: "", data: [] };
}

export const BuildingsResponse: MessageFns<BuildingsResponse> = {
  encode(message: BuildingsResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.isSuccess !== false) {
      writer.uint32(8).bool(message.isSuccess);
    }
    if (message.message !== "") {
      writer.uint32(18).string(message.message);
    }
    if (message.statusCode !== "") {
      writer.uint32(26).string(message.statusCode);
    }
    for (const v of message.data) {
      BuildingsProto.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BuildingsResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBuildingsResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.isSuccess = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.message = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.statusCode = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.data.push(BuildingsProto.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BuildingsResponse {
    return {
      isSuccess: isSet(object.isSuccess) ? globalThis.Boolean(object.isSuccess) : false,
      message: isSet(object.message) ? globalThis.String(object.message) : "",
      statusCode: isSet(object.statusCode) ? globalThis.String(object.statusCode) : "",
      data: globalThis.Array.isArray(object?.data) ? object.data.map((e: any) => BuildingsProto.fromJSON(e)) : [],
    };
  },

  toJSON(message: BuildingsResponse): unknown {
    const obj: any = {};
    if (message.isSuccess !== false) {
      obj.isSuccess = message.isSuccess;
    }
    if (message.message !== "") {
      obj.message = message.message;
    }
    if (message.statusCode !== "") {
      obj.statusCode = message.statusCode;
    }
    if (message.data?.length) {
      obj.data = message.data.map((e) => BuildingsProto.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<BuildingsResponse>, I>>(base?: I): BuildingsResponse {
    return BuildingsResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<BuildingsResponse>, I>>(object: I): BuildingsResponse {
    const message = createBaseBuildingsResponse();
    message.isSuccess = object.isSuccess ?? false;
    message.message = object.message ?? "";
    message.statusCode = object.statusCode ?? "";
    message.data = object.data?.map((e) => BuildingsProto.fromPartial(e)) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}